Move naked fn checks to hir_typeck
